backend/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2-binary==2.9.9
Pillow==8.3.1